apps/server/autocomplete/src/requirements.txt
python-dotenv
openai
aiohttp
PyMuPDF
requests
supabase
tiktoken
fastapi
langchain
langchain-openai
langchain-community
pydantic
tavily-python
PyPDF2
python-multipart
uvicorn
gunicorn
urllib3
baml-py
